requirements.txt
streamlit>=1.20.0
praw>=7.6.0
transformers>=4.30.0
torch
beautifulsoup4
requests
scikit-learn
python-dotenv
vaderSentiment
serpapi

# Added for improved entity matching and NLTK VADER fallback used in sentiment.py
rapidfuzz>=1.9.0
nltk>=3.5
sentencepiece

streamlit>=1.20.0
pandas>=1.3.0
altair>=4.0.0
requests>=2.25.0
